Fri, Apr 8, 2022, 4:05 PM +00:00

+++ readability/readability.c
@@ -15,12 +15,13 @@ int main(void)
     int letters = count_letters(text);
     int words = count_words(text);
     int sentences = count_sentences(text);
+    // calculate grade and store it a variable
     int grade = grade_calculator(letters, words, sentences);
     if (grade < 1)
     {
         printf("Before Grade 1\n");
     }
-    else if(grade >= 16)
+    else if (grade >= 16)
     {
         printf("Grade 16+\n");
     }
@@ -45,7 +46,7 @@ int count_letters(string text)
         // only consider ASCII values that represent uppercase letters
         if (l >= 65 && l <= 90)
         {
-           count += 1;
+            count += 1;
         }
     }
     return count;
@@ -61,7 +62,7 @@ int count_words(string text)
         // count each word by checking spaces
         if (isspace(next_chr))
         {
-            count +=1;
+            count += 1;
         }
         // count the last word of the sentence
         if (i + 1 == strlen(text))
@@ -79,11 +80,12 @@ int count_sentences(string text)
     for (int i = 0; i < strlen(text); i++)
     {
         char chr = text[i];
+        // ascii values for the index
         int ascii = chr;
-        // count each word by checking spaces
+        // for each sentence check if it is: '.','!', or '?'
         if (ascii == 46 || ascii == 63 || ascii == 33)
         {
-            count +=1;
+            count += 1;
         }
     }
     return count;
@@ -94,6 +96,7 @@ int grade_calculator(letters, words, sentences)
     double L = (float)letters / (float)words * 100;
     double S = (float)sentences / (float)words * 100;
     double index = 0.0588 * L - 0.296 * S - 15.8;
-    double rounded_index = round(index);
+    //round index to the nearest integer
+    int rounded_index = round(index);
     return rounded_index;
 }
